+++ crates/astria-sequencer-relayer/src/relayer.rs
@@ -22,7 +22,7 @@ use tracing::{
 
 use crate::{
     base64_string::Base64String,
-    da::CelestiaClient,
+    data_availability::CelestiaClient,
     keys::{
         private_key_bytes_to_keypair,
         validator_hex_to_address,
